Crate webb_primitives

Re-exports

pub use webb_proposals;

pub use hasher::*;

pub use runtime::*;

pub use traits::*;

pub use types::*;

pub use verifier::*;

pub use runtime::*;

Modules

field_ops

hasher

hashing

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

runtime

signing

traits

types

utils

verifier

verifying