Trait webb_primitives::traits::linkable_tree::LinkableTreeInspector

pub trait LinkableTreeInspector<C: LinkableTreeConfig> {
    fn get_chain_id() -> C::ChainId;
    fn get_chain_id_type() -> C::ChainId;
    fn get_chain_type() -> [u8; 2];
    fn is_known_root(
        id: C::TreeId, 
        root: C::Element
    ) -> Result<bool, DispatchError>;
    fn ensure_known_root(
        id: C::TreeId, 
        root: C::Element
    ) -> Result<(), DispatchError>;
    fn get_root(id: C::TreeId) -> Result<C::Element, DispatchError>;
    fn get_neighbor_roots(
        id: C::TreeId
    ) -> Result<Vec<C::Element>, DispatchError>;
    fn is_known_neighbor_root(
        id: C::TreeId, 
        src_chain_id: C::ChainId, 
        target: C::Element
    ) -> Result<bool, DispatchError>;
    fn ensure_known_neighbor_roots(
        id: C::TreeId, 
        roots: &Vec<C::Element>
    ) -> Result<(), DispatchError>;
    fn ensure_known_neighbor_root(
        id: C::TreeId, 
        src_chain_id: C::ChainId, 
        target: C::Element
    ) -> Result<(), DispatchError>;
    fn has_edge(id: C::TreeId, src_chain_id: C::ChainId) -> bool;
    fn ensure_max_edges(
        id: C::TreeId, 
        num_roots: usize
    ) -> Result<(), DispatchError>;
}

Trait for inspecting tree state

Required Methods

fn get_chain_id() -> C::ChainId

fn get_chain_id_type() -> C::ChainId

fn get_chain_type() -> [u8; 2]

fn is_known_root(id: C::TreeId, root: C::Element) -> Result<bool, DispatchError>

Checks if a merkle root is in a tree’s cached history or returns

fn ensure_known_root(
id: C::TreeId,
root: C::Element
) -> Result<(), DispatchError>

fn get_root(id: C::TreeId) -> Result<C::Element, DispatchError>

Gets the merkle root for a tree or returns TreeDoesntExist

fn get_neighbor_roots(id: C::TreeId) -> Result<Vec<C::Element>, DispatchError>

Gets the merkle root for a tree or returns TreeDoesntExist

fn is_known_neighbor_root(
id: C::TreeId,
src_chain_id: C::ChainId,
target: C::Element
) -> Result<bool, DispatchError>

Checks if a merkle root is in a tree’s cached history or returns TreeDoesntExist

fn ensure_known_neighbor_roots(
id: C::TreeId,
roots: &Vec<C::Element>
) -> Result<(), DispatchError>

Checks if each root from passed root array is in tree’s cached history or returns InvalidNeighborWithdrawRoot

fn ensure_known_neighbor_root(
id: C::TreeId,
src_chain_id: C::ChainId,
target: C::Element
) -> Result<(), DispatchError>

Checks if a merkle root is in a tree’s cached history or returns InvalidNeighborWithdrawRoot

fn has_edge(id: C::TreeId, src_chain_id: C::ChainId) -> bool

Check if this linked tree has this edge

fn ensure_max_edges(
id: C::TreeId,
num_roots: usize
) -> Result<(), DispatchError>

Check if passed number of roots is the same as max allowed edges or returns InvalidMerkleRoots