Trait webb_primitives::traits::vanchor::VAnchorInspector

pub trait VAnchorInspector<C: VAnchorConfig> {
    fn is_nullifier_used(id: C::TreeId, nullifier: C::Element) -> bool;
    fn ensure_nullifier_unused(
        id: C::TreeId, 
        nullifier: C::Element
    ) -> Result<(), DispatchError>;
    fn has_edge(id: C::TreeId, src_chain_id: C::ChainId) -> bool;
}

Anchor trait for inspecting tree state

Required Methods

fn is_nullifier_used(id: C::TreeId, nullifier: C::Element) -> bool

Check if a nullifier has been used in a tree or returns InvalidNullifier

fn ensure_nullifier_unused(
id: C::TreeId,
nullifier: C::Element
) -> Result<(), DispatchError>

Check if a nullifier has been used in a tree and throws if not

fn has_edge(id: C::TreeId, src_chain_id: C::ChainId) -> bool

Check if this linked tree has this edge (for backwards compatability)